requirements.txt
# Web 框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# OCR 相关 - 只需要这两个包，其他会自动安装
pyobjc-framework-Vision>=11.1
pillow>=10.0.0

# 系统监控
psutil>=5.9.0

# 日志和工具
python-multipart>=0.0.6
click>=8.0.0

# 开发工具（可选）
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.25.0 